iot-simulator/requirements.txt
requests>=2.32.3
azure-iot-device==2.14.0
